// src/app/models/pokemon.ts
export class Pokemon {
    id: number;
    nombre: string;
    nombre_jp:string;
    url_foto: string;
    url_foto_bk: string;
    url_foto_shiny: string;
    url_foto_shiny_bk: string;
    tipos: string[];
    descripcion: string;
    habilidades: string[];
    alto: number;
    peso: number;
    base_hp: number;
    base_atk: number;
    base_def: number;
    base_spa: number;
    base_spd: number;
    base_spe: number;
}